Delete broken_user test projects and reuse existing
