+++ jib-gradle-plugin/src/integration-test/java/com/google/cloud/tools/jib/gradle/JibRunHelper.java
@@ -75,10 +75,15 @@ public class JibRunHelper {
     assertCreationTimeEpoch(additionalImageReference);
   }
 
-  static void buildToDockerDaemon(TestProject testProject, String imageReference)
+  static void buildToDockerDaemon(
+      TestProject testProject, String imageReference, String gradleBuildFile)
       throws IOException, InterruptedException {
     BuildResult buildResult =
-        testProject.build("clean", "jibDockerBuild", "-D_TARGET_IMAGE=" + imageReference);
+        testProject.build(
+            "clean",
+            "jibDockerBuild",
+            "-D_TARGET_IMAGE=" + imageReference,
+            "-b=" + gradleBuildFile);
     assertBuildSuccess(buildResult, "jibDockerBuild", "Built image to Docker daemon as ");
     Assert.assertThat(buildResult.getOutput(), CoreMatchers.containsString(imageReference));
 
@@ -88,7 +93,7 @@ public class JibRunHelper {
 
   static String buildToDockerDaemonAndRun(TestProject testProject, String imageReference)
       throws IOException, InterruptedException {
-    buildToDockerDaemon(testProject, imageReference);
+    buildToDockerDaemon(testProject, imageReference, "build.gradle");
     return new Command("docker", "run", "--rm", imageReference).run();
   }
