+++ jib-gradle-plugin/src/integration-test/resources/projects/empty/build-broken-user.gradle
