+++ jib-maven-plugin/src/test/resources/projects/empty/pom-broken-user.xml
